Add css file and update code to use UI
Added buttons to take the user input and feed it
to the game functions.

Reworked the getComputer choice function so that
it could be used as a parameter in the game function.

Added an event listener to the wrapper of the new
choice buttons to avoid clutter and wasted memory.

Disabled all choice buttons after the maximum number
of rounds and added a playagain button to make
an overall win or loss clearer.

Fought  java demons and while not victorious, learnt
from our skirmishes!